converted 2d canvas from f32 to f64 and switched coord transform back to cpu

+++ src/graph.rs
@@ -76,7 +76,7 @@ impl Default for DimensionStyle {
 #[derive(Debug, Clone, Copy)]
 pub enum GridSpacing {
     Dynamic { steps: u32, substeps: u32 },
-    Fixed { spacing: f32, substeps: u32 },
+    Fixed { spacing: f64, substeps: u32 },
 }
 
 impl Default for GridSpacing {
